LayoutTests/platform/wk2/http/tests/security/XFrameOptions/x-frame-options-ignore-deny-meta-tag-parent-same-origin-deny-expected.txt

CONSOLE MESSAGE: The X-Frame-Option 'sameorigin' supplied in a <meta> element was ignored. X-Frame-Options may only be provided by an HTTP header sent with the document.
CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: FAIL: Could not read contentWindow.location.href
There should be content in the iframe below



--------
Frame: '<!--frame1-->'
--------
PASS: This should show up even though the parent is not in the same origin because we should be ignoring the meta tag.
